+++ rational/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.1 (WIP)
+
+- Impl `Sum` and `Product` traits for `RBig` and `Relaxed`.
+- Impl `dashu_base::Inverse` trait for `RBig` and `Relaxed`.
+- Add `cubic()` for `RBig` and `Relaxed`.
+
 ## 0.3.0 (Initial release)
 
 - Support basic arithmetic operations and numeric conversion.